+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## v2.0.1 - July 8, 2023
+## v2.0.2 - July 8, 2023
 - Added Spaceship decor
+- Added Setting: Track unhatched seeds
+- Added Setting: Include unhatched seeds in count
 
 ## v2.0.0 - July 4, 2023
 - Adopted Semantic Versioning